SessionGoalPage label now displays time input.

+++ CodingTracker.Common/Entities/CodingSessionEntity.cs
@@ -17,16 +17,16 @@ namespace CodingTracker.Common.Entities.CodingSessionEntities
         public DateTime StartTime { get; set; }
         public DateOnly EndDate { get; set; }
         public DateTime EndTime { get; set; }
-        public double DurationSeconds { get; set; }
+        public int DurationSeconds { get; set; }
 
         public string DurationHHMM { get; set; }
 
         public bool GoalSet { get; set; }
-        public double GoalSeconds { get; set; }
+        public int GoalSeconds { get; set; }
 
         public bool GoalReached { get; set; }
 
-        public string StudyProject { get; set; }
+        public string StudyProject { get; set; } 
         public string StudyNotes { get; set; }
